+++ tasks/remove.js
@@ -4,13 +4,6 @@
 // For more information about Grunt remove, have a look at @openveo/api
 module.exports = {
 
-  // Remove project's compiled documentation
-  doc: {
-    src: [
-      '<%= project.documentationDeployPath %>'
-    ]
-  },
-
   // Remove project's built files and directories
   build: {
     src: [